Created an API for creating a lobby and displaying instance.slug of a new lobby. Replaced the RetrieveDestroyAPIView descendant class with RetrieveAPIView in DetailLobbyView. Reworked the CreateLobbySerializer serializer. Refactored the code (moved the logic of creating boards and ships for boards to the src.game.db_queries module).
